backend/pyproject.toml
[project]
name = "ProfAssistant-Backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "gradio>=5.49.1",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=1.0.0",
    "langsmith>=0.4.37",
    "openai>=2.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
    "python-multipart>=0.0.9",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "chromadb>=1.0.0",
    "wikipedia-mcp>=1.7.0",
    "langchain-mcp-adapters>=0.1.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",
    "bcrypt==4.0.1",
    "aiosqlite>=0.20.0",
    "python-pptx>=1.0.2",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "anyio>=4.11.0",
    "fpdf2>=2.8.0",
    "httpx>=0.28.1",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test_and_evals"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "llm: tests that call a real LLM API (require OPENAI_API_KEY)",
]
